backend/requirements.txt
Flask
flask-cors
gunicorn
python-dotenv
cloudinary
easyocr
opencv-python-headless
google-generativeai==0.8.5
PyMuPDF==1.25.5
torch==2.7.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
